+++ doc/changes.rst
@@ -25,6 +25,9 @@ Deprections
 ==================
 
 
+* `#1047 <https://bitbucket.org/openpyxl/openpyxl/issues/1047>`_ Cannot set quote prefix
+
+
 2.5.6 (2018-08-30)
 ==================